game/requirements.txt
colorama
tk
Pillow
requests
flask